更改bug

+++ app/src/main/java/cn/ifhu/mershop/activity/operation/AddProductActivity.java
@@ -240,7 +240,7 @@ public class AddProductActivity extends BaseActivity {
             return true;
         }
 
-        if (Integer.parseInt(etOriginalPrice.getText().toString().trim()) < Integer.parseInt(etPrice.getText().toString().trim())) {
+        if (Double.parseDouble(etOriginalPrice.getText().toString().trim()) < Double.parseDouble(etPrice.getText().toString().trim())) {
             ToastHelper.makeText("商品原价不能小于商品价格", Toast.LENGTH_SHORT, ToastHelper.NORMALTOAST).show();
             return true;
         }
